markdown: Add support for `![alt text](url)` syntax for inline images.
This commit adds support for standard markdown syntax for images.

Unlike linked image previews (which is the only way images could be
added so far), images added using the `![alt text](url)` syntax are
truly inline, so are placed in the message exactly where the syntax is,
not in a separate paragraph.

Co-authored-by: Nehal Sharma <bablinaneh@gmail.com>

Fixes part of #28912